Create GradesCalculations
Grade Calculator Program

In this program, I created a program which input a csv file from the computer containing all the grades of the semester for an arbitrary student. I wrote this program in C due to the versatility of C. Of course, I could have written this in Python, which would have been easier, however I wanted to understand the theory behind reading in a csv file. In the beginning of this code, I created a "parsing" function which reads the input csv file and understands where the inputs are, such as the name, grades, dates, etc. This was very important because C does not have a function to read in csv file, so it was paramount that I create this parsing function to search the input file. Following that, I created a function to provide the student's grade in the class based on data analysis of the grades.